% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootsmethod.r
\name{bootsmethod}
\alias{bootsmethod}
\title{The SWIRL Sample Size Estimation Method}
\usage{
bootsmethod(data, dist, bmrkpar, ...)
}
\arguments{
\item{data}{is a data frame containing the data set}

\item{dist}{is for the distribution of the biomarker. Specify either normal or uniform}

\item{bmrkpar}{is a vector containing the parameters of the distribution.}
}
\value{
a list containing values for Theta, 95\% interval for Theta and the converage probability for Theta
}
\description{
This function is used to estimate the sample size that guarantee the
95\% CI of the paramter Theta to be within user defined width. It first
generate a data set from the model used to calculate Theta with a sample size ranging from
300 to 3500 and calculate the 95|% CI width of each sample. Using OLS , then a model is fitted
with n as response and W as a covariate.
}
