% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendatlog.R
\name{coeffgen}
\alias{coeffgen}
\title{Estimating the modparicients For the Logistic Regression MOdel}
\usage{
coeffgen(k, z)
}
\arguments{
\item{k}{a vector of four user defined elements}

\item{z}{a vector of two elements, the 25th and 75th percetile value of the biomarker}
}
\value{
a vector of beta modparicients
}
\description{
This Function is used to Generate the beta modparicients from the
user specified 'K' values. These Beta are then used to generate data
from the logistic regression model.
}
\details{
This function take the values of K1-k4 and Z1-Z2 and return the four beta values.
The Beta values correspond to the modparicients in the multiple logistic regression
we fitted. The first coef is for the intercept, the 2nd for the modpar of the biomarker
, the 3rd for the treatment and the last for the interaction term.
}
\author{
Henok Woldu
}
