% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendatlog.R
\name{logdata}
\alias{logdata}
\title{Generating Data From Multiple Logistic Regression Model}
\usage{
logdata(m, n, modpar, dist, bmrkpar, bmrkdgn = "stratified", ...)
}
\arguments{
\item{m}{is the number of data set needed to be generated}

\item{n}{is the number of sample size needed}

\item{modpar}{is the vector of the beta model parameters (coefficients)}

\item{dist}{is the distribution of the biomarker}

\item{bmrkpar}{parameters of the biomarker distribution}

\item{bmrkdgn}{is the type of biomarker study design}
}
\value{
a list with 'm' data set each having 'n' sample size
}
\description{
This function is used to generate data from a multiple logistic
regression model using the modparicients obtained from the \code{Betas}
function.
}
