% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWIRL.r
\name{ssswirl}
\alias{ssswirl}
\title{Sample Size Using SWIRL Method}
\usage{
ssswirl(modpar, dist, mcdat, m, bmrkpar, new, ...)
}
\arguments{
\item{modpar}{a vector containing the model coefficients}

\item{dist}{is for the distribution of the biomarker. Specify normal or uniform}

\item{mcdat}{is a vector of data with sample size arranged from the smallest to the largest}

\item{m}{number of data sets to generate for a given sample size}

\item{bmrkpar}{parameters for the distribution of the biomarker.}

\item{new}{is a vector for user specified width.}
}
\value{
a numeric value of estimated sample size for a given used specificed width.
}
\description{
This function is used to estimate the sample size needed to guarantee the
95\% CI of the parameter theta is within used defined width.
}
